[FIX] cetmix_tower_server: Clean up Python command execution
This commit fixes the following issue:

Expected result
---------------
The Python command does not necessarily have to execute successfully, but the secret values ​​must be replaced with the correct placeholders in the output.

Current result
--------------
The Python command does not necessarily have to execute successfully, but secret values are not handled or replaced correctly in the output.

Solution
--------
Renamed method _parse_ssh_command_results to _parse_command_results for using not only for SSH but for any command including Python one.

Task: 3952