[IMP] cetmix_tower_yaml: Secret value check
When exporting a key the secret value is replaced with the mask.
This is done to prevent leaking sensitive data during export.
However this may create an issue when the same file is imported back.
Because if this mask still remains in the imported file
it will overwrite the actual secret value.

This commit prevents this from happening by checking the secret value
before creating or updating a key.
If the mask is still present in the file as a value an error is raised.

Task: 4119